cache split files in folders
