+++ lib/cache/CheckOptions.ts
